requirements.txt
python-dotenv
sqlalchemy
twitchio
bottle
pyyaml